YearMonthUtilを別ファイルに外だし。
